Fix some threading issues in IconDatabase. https://bugs.webkit.org/show_bug.cgi?id=57985
Reviewed by Brady Eidson.

No functionality change so no new tests. It would be very hard to test this, but
I am working on a change for bug 31639 to detect these (which makes layout tests
fail for at least one of these issues).

* loader/icon/IconDatabase.cpp:
(WebCore::IconDatabase::readFromDatabase): Release imageData so that it won't get deref'ed outside of locks.
(WebCore::IconDatabase::writeToDatabase): Acquire m_urlAndIconLock because the SharedBuffer<> data
is used, and this lock seems to be the one which consistently guards it.

Identifier: 73667@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@83913 268f45cc-cd09-0410-ab3c-d52691b4dbfc