Hang underneath ApplicationCacheStorage::writeDataToUniqueFileInDirectory when loading http://www.webkit.org/demos/calendar. https://bugs.webkit.org/show_bug.cgi?id=58698
Reviewed by Brian Weinstein.

* platform/win/FileSystemWin.cpp:
(WebCore::openFile):
Add breaks to the case statement.
(WebCore::directoryName):
Remove any trailing slash in directoryName. After closer inspection of all callers, it
appears the that callers do not need the trailing slash, and those that use the length do
so only to change the direction of the slashes throughout the length of the path.



Identifier: 73792@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84051 268f45cc-cd09-0410-ab3c-d52691b4dbfc