https://bugs.webkit.org/show_bug.cgi?id=57898 REGRESSION (r82185): Scroll position not restored on navigation back to a page in the page cache -and corresponding- <rdar://problem/9226652>
Reviewed by Maciej Stachowiak.

Source/WebCore:

Setting the contentsSize of a ScrollView to (0, 0) necessarily causes the scroll
position to be lost. (The scroll position is computed based on the
visibleContentSize.) This patch provides a mechanism to cache the current scroll
position, and then the HistoryController accesses only the cached position rather
than calling scrollPosition() which does a computation based on the
visibleContentSize.

* loader/HistoryController.cpp:
(WebCore::HistoryController::saveScrollPositionAndViewStateToItem):
* page/FrameView.cpp:
(WebCore::FrameView::resetScrollbarsAndClearContentsSize):
* platform/ScrollView.h:
(WebCore::ScrollView::cacheCurrentScrollPosition):
(WebCore::ScrollView::cachedScrollPosition):

LayoutTests:

* fast/loader/resources/empty-document-goes-back.html: Added.
* fast/loader/scroll-position-restored-on-back-expected.txt: Added.
* fast/loader/scroll-position-restored-on-back.html: Added.



Identifier: 74036@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84296 268f45cc-cd09-0410-ab3c-d52691b4dbfc