<rdar://problem/8891395> -[DOMRange textRects] returns incorrect results for vertical or flipped text https://bugs.webkit.org/show_bug.cgi?id=60067
Reviewed by Darin Adler.

No test because this code path is only used by the Objective-C API, which is not testable from DumpRenderTree.

* rendering/RenderText.cpp:
(WebCore::RenderText::absoluteRectsForRange): Use width/height instead of logicalWidth/logicalHeight here,
and perform the local-to-absolute mapping on the rects rather than their origin, in order to get the right
results for flipped writing modes.



Identifier: 75364@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@85677 268f45cc-cd09-0410-ab3c-d52691b4dbfc