Source/WebCore: WebKit2: Docking Web Inspector doesn't respect maximum inspector size https://bugs.webkit.org/show_bug.cgi?id=60294 <rdar://problem/9388518>
Reviewed by Timothy Hatcher.

Make a needed function public and export it.

* WebCore.exp.in: Export constrainedAttaechedWindowHeight.
* inspector/InspectorFrontendClientLocal.h: Make constrainedAttaechedWindowHeight public.

Source/WebKit2: WebKit2: Docking Web Inspector doesn't respect maximum inspector size
https://bugs.webkit.org/show_bug.cgi?id=60294
<rdar://problem/9388518>

Reviewed by Timothy Hatcher.

Before we place the inspector and the inspected views, make sure that we have constrained the height
of the docked inspector. Putting the code here covers both when the inspector is initially attached
and when a window with the docked inspector is resized.

* UIProcess/mac/WebInspectorProxyMac.mm:
(WebKit::WebInspectorProxy::inspectedViewFrameDidChange): Constrain the height before laying out the views.
* UIProcess/win/WebInspectorProxyWin.cpp:
(WebKit::WebInspectorProxy::onWebViewWindowPosChangingEvent): Ditto.



Identifier: 75557@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@85874 268f45cc-cd09-0410-ab3c-d52691b4dbfc