<rdar://problem/9446653> REGRESSION (r84750): Moving by word stops at apostrophe mid-word https://bugs.webkit.org/show_bug.cgi?id=60915
Reviewed by Darin Adler.

Source/WebCore:

CFStringTokenizer’s kCFStringTokenizerUnitWord considers “Here’s” as two separate tokens.
Switching to CFStringTokenizer in r84750 was an attempt to address an issue with Japanese word
boundaries for searches with WebFindOptionsAtWordStarts, but it turned out to be insufficient,
and in r86387 the Japanese word issue was addressed independently of text boundaries, so just
revert r84750.

* platform/text/mac/TextBoundaries.mm:
(WebCore::findNextWordFromIndex):

LayoutTests:

* editing/text-iterator/findString-expected.txt:
* editing/text-iterator/findString.html:



Identifier: 76249@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@86599 268f45cc-cd09-0410-ab3c-d52691b4dbfc