https://bugs.webkit.org/show_bug.cgi?id=62082 DFG JIT - bug passing arguments that need swap
Reviewed by Sam Weinig.

This is really just a typo.
When setting up the arguments for a call out to a C operation, we'll
fail to swap arguments where this is necessary. For example, in the
case of 2 arg calls, where the first argument is in %rdx & the second
is in %rsi we should swap (exec will be passed in %rdi), but we don't.

This can also affect function calls passing three arguments.

* dfg/DFGJITCodeGenerator.h:
(JSC::DFG::JITCodeGenerator::setupTwoStubArgs):
    - Call swap with the correct arguments.



Identifier: 77544@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@88092 268f45cc-cd09-0410-ab3c-d52691b4dbfc