https://bugs.webkit.org/show_bug.cgi?id=63669 DFG JIT - fix spectral-norm regression
Reviewed by Geoff Garen.

The problem is a mis-speculation leading to us falling off the speculative path.
Make the speculation logic slightly smarter, don't predict int if one of the
operands is already loaded as a double (we use this logic already for compares).

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::compile):
* dfg/DFGSpeculativeJIT.h:
(JSC::DFG::SpeculativeJIT::shouldSpeculateInteger):



Identifier: 79346@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@90065 268f45cc-cd09-0410-ab3c-d52691b4dbfc