Teach garden-o-matic how to display test results https://bugs.webkit.org/show_bug.cgi?id=64141
Reviewed by Ojan Vafai.

This patch includes basic infrastructure for probing build.chromium.org
for test results.  We only handle text and image tests, not anything
complicated like reftests.  Also, we're using the revision/build
independent results store on the server, so we're avoiding that
complication for now.

It's slightly hacky that we need to probe the server to see what kinds
of results exist.  A better solution would be to add CORS support to
the server or to use the local server to help.

* Scripts/webkitpy/tool/servers/data/gardeningserver/base.js:
* Scripts/webkitpy/tool/servers/data/gardeningserver/index.html:
* Scripts/webkitpy/tool/servers/data/gardeningserver/main.js:
* Scripts/webkitpy/tool/servers/data/gardeningserver/results.js:
* Scripts/webkitpy/tool/servers/data/gardeningserver/ui.js:
* Scripts/webkitpy/tool/servers/data/gardeningserver/ui_unittests.js:


Identifier: 79917@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@90652 268f45cc-cd09-0410-ab3c-d52691b4dbfc